requirements.txt
telethon==1.34.0
python-dotenv==1.0.1
PyYAML==6.0.1
pytz==2024.1

pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2 